feat: implement initial functions trigger
