+++ package.json
@@ -48,6 +48,7 @@
   "license": "MIT",
   "dependencies": {
     "@azure/cosmos": "^3.17.3",
+    "@azure/functions": "^3.5.1",
     "@elastic/elasticsearch": "7.12",
     "@types/aws-lambda": "^8.10.62",
     "@types/express": "^4.17.11",
